Renamed validate score period info function and moved it into score period modal general file
